Facial-Attendance/requirements.txt
altgraph==0.17.4
click==8.1.7
colorama==0.4.6
dlib==19.22.99
face-recognition==1.3.0
face_recognition_models==0.3.0
importlib_metadata==8.4.0
mysql-connector-python==9.0.0
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pywin32-ctypes==0.2.3
scipy==1.10.1
zipp==3.20.1
